fix TypeError: function takes exactly 3 arguments (2 given) from wait_for_edge
Trying to parse 3 required arguments and then falling back to 2 will set the
PyError condition when only two arguments are present even though the
wait_for_edge function as a whole succeeds.  According to
https://docs.python.org/2/c-api/exceptions.html, returning normally after this
will lead to undefined behavior the next time the code enters the C API.

Fix this by using the PyArg_ParseTuple optional argument feature instead of
trying to parse 3 arguments and falling back to 2 if it fails.